Added tests for AddressServiceImpl
