+++ service/src/test/java/greencity/service/ubs/UBSClientServiceImplTest.java
@@ -21,8 +21,6 @@ import greencity.dto.courier.CourierDto;
 import greencity.dto.customer.UbsCustomersDto;
 import greencity.dto.customer.UbsCustomersDtoUpdate;
 import greencity.dto.employee.UserEmployeeAuthorityDto;
-import greencity.dto.location.api.DistrictDto;
-import greencity.dto.location.api.LocationDto;
 import greencity.dto.order.EventDto;
 import greencity.dto.order.OrderAddressDtoRequest;
 import greencity.dto.order.OrderCancellationReasonDto;
@@ -62,8 +60,6 @@ import greencity.entity.telegram.TelegramBot;
 import greencity.entity.user.Location;
 import greencity.entity.user.User;
 import greencity.entity.user.employee.Employee;
-import greencity.entity.user.locations.City;
-import greencity.entity.user.locations.District;
 import greencity.entity.user.ubs.Address;
 import greencity.entity.user.ubs.OrderAddress;
 import greencity.entity.user.ubs.UBSuser;
@@ -84,7 +80,6 @@ import greencity.exceptions.address.AddressNotWithinLocationAreaException;
 import greencity.exceptions.http.AccessDeniedException;
 import greencity.exceptions.user.UBSuserNotFoundException;
 import greencity.exceptions.user.UserNotFoundException;
-import greencity.mapping.location.AddressRequestDtoToBaseEntityMapper;
 import greencity.mapping.location.LocationToLocationsDtoMapper;
 import greencity.repository.AddressRepository;
 import greencity.repository.BagRepository;
@@ -119,7 +114,6 @@ import greencity.util.EncryptionUtil;
 import greencity.util.OrderUtils;
 import jakarta.persistence.EntityNotFoundException;
 import java.lang.reflect.Field;
-import java.lang.reflect.Method;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -162,7 +156,6 @@ import static greencity.ModelUtils.TEST_EMAIL;
 import static greencity.ModelUtils.TEST_ORDER_ADDRESS_DTO_UPDATE;
 import static greencity.ModelUtils.TEST_PAYMENT_LIST;
 import static greencity.ModelUtils.TEST_UUID;
-import static greencity.ModelUtils.addressDto;
 import static greencity.ModelUtils.addressDtoList;
 import static greencity.ModelUtils.addressDtoListWithNullPlaceId;
 import static greencity.ModelUtils.addressList;
@@ -172,15 +165,6 @@ import static greencity.ModelUtils.botList;
 import static greencity.ModelUtils.createCertificateDto;
 import static greencity.ModelUtils.getActiveCertificateWith10Points;
 import static greencity.ModelUtils.getAddress;
-import static greencity.ModelUtils.getAddressDtoResponse;
-import static greencity.ModelUtils.getAddressRequestDto;
-import static greencity.ModelUtils.getAddressRequestDto2;
-import static greencity.ModelUtils.getAddressRequestDtoReflection;
-import static greencity.ModelUtils.getAddressRequestDtoReflection2;
-import static greencity.ModelUtils.getAddressRequestDtoReflection3;
-import static greencity.ModelUtils.getAddressRequestDtoReflection4;
-import static greencity.ModelUtils.getAddressRequestDtoReflection5;
-import static greencity.ModelUtils.getAddressRequestToSaveDto;
 import static greencity.ModelUtils.getAddressWithKyivRegionToSaveRequestDto;
 import static greencity.ModelUtils.getBag;
 import static greencity.ModelUtils.getBag1list;
@@ -199,7 +183,6 @@ import static greencity.ModelUtils.getEvent1;
 import static greencity.ModelUtils.getEvent2;
 import static greencity.ModelUtils.getGeocodingResultWithKyivRegion;
 import static greencity.ModelUtils.getLocation;
-import static greencity.ModelUtils.getMaximumAmountOfAddresses;
 import static greencity.ModelUtils.getMonoBankPaymentResponseDto;
 import static greencity.ModelUtils.getNotificationPaymentLink;
 import static greencity.ModelUtils.getOrder;
@@ -230,9 +213,7 @@ import static greencity.ModelUtils.getTariffsForLocationDto;
 import static greencity.ModelUtils.getTariffsInfo;
 import static greencity.ModelUtils.getTelegramBotNotifyTrue;
 import static greencity.ModelUtils.getTestOrderAddressDtoRequest;
-import static greencity.ModelUtils.getTestOrderAddressDtoRequest2;
 import static greencity.ModelUtils.getTestOrderAddressDtoRequestWithNullPlaceId;
-import static greencity.ModelUtils.getTestOrderAddressLocationDto;
 import static greencity.ModelUtils.getTestUser;
 import static greencity.ModelUtils.getUBSuser;
 import static greencity.ModelUtils.getUBSuserWithoutSender;
@@ -251,15 +232,8 @@ import static greencity.ModelUtils.getUserWithBotNotifyTrue;
 import static greencity.ModelUtils.getUserWithLastLocation;
 import static greencity.ModelUtils.getViberBotNotifyTrue;
 import static greencity.constant.AppConstant.USER_WITH_PREFIX;
-import static greencity.constant.ErrorMessage.ACTUAL_ADDRESS_NOT_FOUND;
-import static greencity.constant.ErrorMessage.ADDRESS_ALREADY_EXISTS;
-import static greencity.constant.ErrorMessage.CANNOT_ACCESS_PERSONAL_INFO;
-import static greencity.constant.ErrorMessage.CANNOT_DELETE_ADDRESS;
-import static greencity.constant.ErrorMessage.CANNOT_DELETE_ALREADY_DELETED_ADDRESS;
-import static greencity.constant.ErrorMessage.CANNOT_MAKE_ACTUAL_DELETED_ADDRESS;
 import static greencity.constant.ErrorMessage.LOCATION_DOESNT_FOUND_BY_ID;
 import static greencity.constant.ErrorMessage.LOCATION_IS_DEACTIVATED_FOR_TARIFF;
-import static greencity.constant.ErrorMessage.NOT_FOUND_ADDRESS_ID_FOR_CURRENT_USER;
 import static greencity.constant.ErrorMessage.ORDER_DOES_NOT_BELONG_TO_USER;
 import static greencity.constant.ErrorMessage.ORDER_WITH_CURRENT_ID_DOES_NOT_EXIST;
 import static greencity.constant.ErrorMessage.PAYMENT_VALIDATION_ERROR;
@@ -420,32 +394,10 @@ class UBSClientServiceImplTest {
     private NotificationParameterRepository notificationParameterRepository;
 
     @Mock
-    private AddressRequestDtoToBaseEntityMapper addressMapper;
-
-    @Mock
     private AddressService addressService;
 
     @Value("${greencity.monobank.token}")
     private String token;
-    private static final String USER_UUID = "uuid";
-
-    @Test
-    void testGetAllDistricts() {
-
-        List<LocationDto> locationDtos;
-        List<DistrictDto> districtDtos;
-        locationDtos = Arrays.asList(LocationDto.builder().build(), LocationDto.builder().build());
-        districtDtos = Arrays.asList(DistrictDto.builder().build(), DistrictDto.builder().build());
-
-        when(locationApiService.getAllDistrictsInCityByNames(anyString(), anyString())).thenReturn(locationDtos);
-        when(modelMapper.map(any(LocationDto.class), eq(DistrictDto.class))).thenAnswer(i -> new DistrictDto());
-
-        List<DistrictDto> results = addressService.getAllDistricts("region", "city");
-
-        assertEquals(districtDtos.size(), results.size());
-        verify(locationApiService, times(1)).getAllDistrictsInCityByNames(anyString(), anyString());
-        verify(modelMapper, times(locationDtos.size())).map(any(LocationDto.class), eq(DistrictDto.class));
-    }
 
     @Test
     void getFirstPageDataByTariffAndLocationIdShouldThrowExceptionWhenTariffLocationDoesNotExist() {
@@ -2037,105 +1989,6 @@ class UBSClientServiceImplTest {
     }
 
     @Test
-    void testFindAllAddressesForCurrentOrder() {
-        String uuid = "35467585763t4sfgchjfuyetf";
-        User user = new User();
-        user.setId(13L);
-        when(userRepository.findByUuid(uuid)).thenReturn(user);
-
-        List<AddressDto> testAddressesDto = getTestAddressesDto();
-
-        OrderWithAddressesResponseDto expected = new OrderWithAddressesResponseDto(testAddressesDto);
-
-        List<Address> addresses = getTestAddresses(user);
-
-        when(addressRepository.findAllNonDeletedAddressesByUserId(user.getId())).thenReturn(addresses);
-        when(modelMapper.map(addresses.get(0), AddressDto.class)).thenReturn(testAddressesDto.get(0));
-        when(modelMapper.map(addresses.get(1), AddressDto.class)).thenReturn(testAddressesDto.get(1));
-
-        OrderWithAddressesResponseDto actual = addressService.findAllAddressesForCurrentOrder(uuid);
-
-        assertEquals(actual, expected);
-        verify(userRepository, times(1)).findByUuid(uuid);
-        verify(addressRepository, times(1)).findAllNonDeletedAddressesByUserId(user.getId());
-    }
-
-    private List<Address> getTestAddresses(User user) {
-        Address address1 = Address.builder()
-            .addressStatus(AddressStatus.NEW).id(13L).city("Kyiv").district("Svyatoshyn")
-            .entranceNumber("1").houseCorpus("1").houseNumber("55").street("Peremohy av.")
-            .user(user).actual(true).coordinates(new Coordinates(12.5, 34.5))
-            .build();
-
-        Address address2 = Address.builder()
-            .addressStatus(AddressStatus.NEW).id(42L).city("Lviv").district("Syhiv")
-            .entranceNumber("1").houseCorpus("1").houseNumber("55").street("Lvivska st.")
-            .user(user).actual(true).coordinates(new Coordinates(13.5, 36.5))
-            .build();
-
-        return Arrays.asList(address1, address2);
-    }
-
-    private List<AddressDto> getTestAddressesDto() {
-        AddressDto addressDto1 = AddressDto.builder().actual(true).id(13L).city("Kyiv").district("Svyatoshyn")
-            .entranceNumber("1").houseCorpus("1").houseNumber("55").street("Peremohy av.")
-            .coordinates(new Coordinates(12.5, 34.5)).build();
-
-        AddressDto addressDto2 = AddressDto.builder().actual(true).id(42L).city("Lviv").district("Syhiv")
-            .entranceNumber("1").houseCorpus("1").houseNumber("55").street("Lvivska st.")
-            .coordinates(new Coordinates(13.5, 36.5)).build();
-        return Arrays.asList(addressDto1, addressDto2);
-    }
-
-    @Test
-    void testSaveCurrentAddressForOrder() {
-        User user = getUserForCreate();
-        List<Address> addresses = user.getAddresses();
-        addresses.getFirst().setActual(false);
-        addresses.getFirst().setAddressStatus(AddressStatus.NEW);
-
-        String uuid = user.getUuid();
-        CreateAddressRequestDto createAddressRequestToSaveDto = getAddressRequestToSaveDto();
-        Address addressToSave = getAddress();
-
-        when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
-        when(addressRepository.findAllNonDeletedAddressesByUserId(user.getId())).thenReturn(addresses);
-
-        when(modelMapper.map(any(), eq(Address.class))).thenReturn(addressToSave);
-        when(regionRepository.findRegionByNameEnOrNameUk(any(), any())).thenReturn(Optional.of(getRegion()));
-        when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
-            .thenReturn(Optional.of(getCity()));
-        when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
-            .thenReturn(Optional.of(getDistrict()));
-
-        var addressDto = addressDto();
-        addressDto.setDistrict("Район");
-        addressDto.setDistrictEn("District");
-        when(modelMapper.map(addresses.getFirst(), AddressDto.class)).thenReturn(addressDto);
-
-        OrderWithAddressesResponseDto actualWithSearchAddress =
-                addressService.saveCurrentAddressForOrder(createAddressRequestToSaveDto, uuid);
-
-        assertEquals(getAddressDtoResponse(), actualWithSearchAddress);
-        assertEquals(createAddressRequestToSaveDto.getDistrict(),
-            actualWithSearchAddress.getAddressList().getFirst().getDistrict());
-        assertEquals(createAddressRequestToSaveDto.getDistrictEn(),
-            actualWithSearchAddress.getAddressList().getFirst().getDistrictEn());
-
-        verify(addressRepository).save(addressToSave);
-
-        verify(userRepository, times(2)).findByUuid(user.getUuid());
-        verify(addressRepository, times(2)).findAllNonDeletedAddressesByUserId(user.getId());
-        verify(regionRepository).findRegionByNameEnOrNameUk(anyString(), anyString());
-        verify(cityRepository).findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString());
-        verify(districtRepository).findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString());
-
-        verify(modelMapper, times(1)).map(any(), eq(Address.class));
-        verify(modelMapper).map(any(), eq(Address.class));
-        verify(modelMapper).map(addresses.getFirst(), AddressDto.class);
-    }
-
-    @Test
     void saveCurrentAddressForOrderForAddressesBelongToKyivEnTest() {
         User user = getUserForCreate();
         List<Address> addresses = user.getAddresses();
@@ -2213,415 +2066,6 @@ class UBSClientServiceImplTest {
     }
 
     @Test
-    void testSaveCurrentAddressForOrderAlreadyExistException() {
-        User user = getUserForCreate();
-        List<Address> addresses = user.getAddresses();
-        String uuid = user.getUuid();
-        OrderAddressDtoRequest dtoRequest = getTestOrderAddressLocationDto();
-
-        CreateAddressRequestDto createAddressRequestDto = getAddressRequestDto();
-
-        addresses.getFirst().setAddressStatus(AddressStatus.NEW);
-        addresses.getFirst().setActual(false);
-
-        when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
-        when(addressRepository.findAllNonDeletedAddressesByUserId(user.getId())).thenReturn(addresses);
-        when(modelMapper.map(any(),
-            eq(CreateAddressRequestDto.class)))
-            .thenReturn(dtoRequest);
-        when(addressRepository.findAllByUserId(user.getId())).thenReturn(addresses);
-
-        dtoRequest.setPlaceId(null);
-
-        BadRequestException exception = assertThrows(BadRequestException.class,
-            () -> addressService.saveCurrentAddressForOrder(createAddressRequestDto, uuid));
-
-        assertEquals(ADDRESS_ALREADY_EXISTS, exception.getMessage());
-
-        verify(userRepository).findByUuid(user.getUuid());
-        verify(addressRepository).findAllNonDeletedAddressesByUserId(user.getId());
-        verify(modelMapper, times(2)).map(any(), eq(CreateAddressRequestDto.class));
-    }
-
-    @Test
-    void testSaveCurrentAddressForMaximumNumbersOfOrdersAddressesException() {
-        User user = getUserForCreate();
-        List<Address> addresses = getMaximumAmountOfAddresses();
-        String uuid = user.getUuid();
-        CreateAddressRequestDto createAddressRequestDto = getAddressRequestDto();
-
-        when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
-        when(addressRepository.findAllNonDeletedAddressesByUserId(user.getId())).thenReturn(addresses);
-
-        BadRequestException exception = assertThrows(BadRequestException.class,
-            () -> addressService.saveCurrentAddressForOrder(createAddressRequestDto, uuid));
-
-        assertEquals(ErrorMessage.NUMBER_OF_ADDRESSES_EXCEEDED, exception.getMessage());
-    }
-
-    @Test
-    void testUpdateCurrentAddressForOrder() {
-        User user = getUserForCreate();
-        List<Address> addresses = user.getAddresses();
-        String uuid = user.getUuid();
-        OrderAddressDtoRequest dtoRequest = getTestOrderAddressLocationDto();
-        dtoRequest.setId(1L);
-        OrderAddressDtoRequest updateAddressRequestDto = getTestOrderAddressDtoRequest();
-        updateAddressRequestDto.setId(1L);
-        addresses.getFirst().setActual(false);
-        addresses.getFirst().setAddressStatus(AddressStatus.IN_ORDER);
-        addresses.getFirst().setUser(user);
-
-        when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
-        when(addressRepository.findAllNonDeletedAddressesByUserId(user.getId())).thenReturn(addresses);
-        when(addressRepository.findById(updateAddressRequestDto.getId()))
-            .thenReturn(Optional.ofNullable(addresses.getFirst()));
-        when(modelMapper.map(any(),
-            eq(Address.class))).thenReturn(addresses.getFirst());
-        when(addressRepository.save(addresses.getFirst())).thenReturn(addresses.getFirst());
-        when(regionRepository.findRegionByNameEnOrNameUk(any(), any())).thenReturn(Optional.of(getRegion()));
-        when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
-            .thenReturn(Optional.of(getCity()));
-        when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
-            .thenReturn(Optional.of(getDistrict()));
-
-        var addressDto = addressDto();
-        addressDto.setDistrict("Район");
-        addressDto.setDistrictEn("District");
-
-        when(modelMapper.map(addresses.getFirst(),
-            AddressDto.class))
-            .thenReturn(addressDto);
-
-        OrderWithAddressesResponseDto actualWithSearchAddress =
-                addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid);
-
-        assertEquals(getAddressDtoResponse(), actualWithSearchAddress);
-        assertEquals(updateAddressRequestDto.getDistrict(),
-            actualWithSearchAddress.getAddressList().getFirst().getDistrict());
-        assertEquals(updateAddressRequestDto.getDistrictEn(),
-            actualWithSearchAddress.getAddressList().getFirst().getDistrictEn());
-
-        OrderWithAddressesResponseDto actualWithoutSearchAddress =
-                addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid);
-
-        assertEquals(getAddressDtoResponse(), actualWithoutSearchAddress);
-
-        verify(addressRepository, times(2)).save(addresses.getFirst());
-        verify(regionRepository, times(2)).findRegionByNameEnOrNameUk(anyString(), anyString());
-        verify(cityRepository, times(2)).findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString());
-        verify(districtRepository, times(2)).findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString());
-    }
-
-    @Test
-    void testUpdateCurrentAddressForOrderWithNoAddress() {
-        User user = getUserForCreate(AddressStatus.DELETED);
-        List<Address> addresses = user.getAddresses();
-        String uuid = user.getUuid();
-        OrderAddressDtoRequest dtoRequest = getTestOrderAddressLocationDto();
-        dtoRequest.setId(7L);
-        OrderAddressDtoRequest updateAddressRequestDto = getTestOrderAddressDtoRequest();
-        updateAddressRequestDto.setId(1L);
-        addresses.getFirst().setActual(false);
-        addresses.getFirst().setUser(user);
-
-        when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
-        when(addressRepository.findById(updateAddressRequestDto.getId()))
-            .thenReturn(Optional.ofNullable(addresses.getFirst()));
-        when(addressRepository.findById(updateAddressRequestDto.getId()))
-            .thenReturn(Optional.ofNullable(addresses.getFirst()));
-        when(regionRepository.findRegionByNameEnOrNameUk(any(), any())).thenReturn(Optional.of(getRegion()));
-        when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
-            .thenReturn(Optional.of(getCity()));
-        when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
-            .thenReturn(Optional.of(getDistrict()));
-        when(modelMapper.map(any(),
-            eq(Address.class))).thenReturn(addresses.getFirst());
-
-        when(addressRepository.save(addresses.getFirst())).thenReturn(addresses.getFirst());
-
-        OrderWithAddressesResponseDto actualWithSearchAddress =
-                addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid);
-
-        assertEquals(OrderWithAddressesResponseDto.builder().addressList(Collections.emptyList()).build(),
-            actualWithSearchAddress);
-
-        verify(addressRepository, times(1)).save(addresses.getFirst());
-        verify(addressRepository, times(1)).findById(anyLong());
-        verify(modelMapper, times(1)).map(any(), eq(Address.class));
-    }
-
-    @Test
-    void testUpdateCurrentAddressForOrderAlreadyExistException() {
-        User user = getUserForCreate();
-        List<Address> addresses = user.getAddresses();
-        String uuid = user.getUuid();
-        OrderAddressDtoRequest dtoRequest = getTestOrderAddressLocationDto();
-        dtoRequest.setId(1L);
-        dtoRequest.setPlaceId(null);
-        OrderAddressDtoRequest updateAddressRequestDto = getTestOrderAddressDtoRequest();
-        updateAddressRequestDto.setId(1L);
-        addresses.getFirst().setActual(false);
-        addresses.getFirst().setAddressStatus(AddressStatus.IN_ORDER);
-        addresses.getFirst().setUser(user);
-
-        when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
-        when(addressRepository.findById(updateAddressRequestDto.getId()))
-            .thenReturn(Optional.ofNullable(addresses.getFirst()));
-        when(modelMapper.map(any(),
-            eq(CreateAddressRequestDto.class)))
-            .thenReturn(dtoRequest);
-        when(addressRepository.findAllByUserId(user.getId())).thenReturn(addresses);
-
-        BadRequestException exception = assertThrows(BadRequestException.class,
-            () -> addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid));
-
-        assertEquals(ADDRESS_ALREADY_EXISTS, exception.getMessage());
-    }
-
-    @Test
-    void testUpdateCurrentAddressForOrderNotFoundOrderAddressException() {
-        User user = getUserForCreate();
-        List<Address> addresses = user.getAddresses();
-        String uuid = user.getUuid();
-        OrderAddressDtoRequest dtoRequest = getTestOrderAddressLocationDto();
-        dtoRequest.setId(1L);
-        OrderAddressDtoRequest updateAddressRequestDto = getTestOrderAddressDtoRequest();
-        updateAddressRequestDto.setId(1L);
-        addresses.getFirst().setActual(false);
-        addresses.getFirst().setAddressStatus(AddressStatus.IN_ORDER);
-        addresses.getFirst().setUser(user);
-
-        when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
-        when(addressRepository.findById(updateAddressRequestDto.getId())).thenReturn(Optional.empty());
-
-        NotFoundException exception = assertThrows(NotFoundException.class,
-            () -> addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid));
-
-        assertEquals(NOT_FOUND_ADDRESS_ID_FOR_CURRENT_USER + updateAddressRequestDto.getId(), exception.getMessage());
-    }
-
-    @Test
-    void testUpdateCurrentAddressForOrderThrowsAccessDeniedException() {
-        long addressId = 1L;
-        long userId = 2L;
-
-        User user = getUserForCreate();
-        user.setId(userId);
-
-        Address address = new Address();
-        address.setId(addressId);
-        address.setUser(new User());
-        address.getUser().setId(userId + 1);
-
-        String uuid = user.getUuid();
-        OrderAddressDtoRequest dtoRequest = getTestOrderAddressLocationDto();
-        dtoRequest.setId(addressId);
-
-        when(addressRepository.findById(addressId)).thenReturn(Optional.of(address));
-        when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
-
-        AccessDeniedException exception = assertThrows(AccessDeniedException.class,
-            () -> addressService.updateCurrentAddressForOrder(dtoRequest, uuid));
-
-        assertEquals(CANNOT_ACCESS_PERSONAL_INFO, exception.getMessage());
-    }
-
-    @Test
-    void testDeleteCurrentAddressForOrderWhenAddressIsActual() {
-        UBSClientServiceImpl ubsClientServiceSpy = spy(ubsService);
-
-        Long firstAddressId = 1L;
-        Long secondAddressId = 2L;
-        Address firstAddress = getAddress();
-        firstAddress.setId(firstAddressId);
-        firstAddress.setActual(true);
-        Address secondAddress = getAddress();
-        secondAddress.setId(secondAddressId);
-        User user = getUser();
-        firstAddress.setUser(user);
-        String uuid = user.getUuid();
-
-        when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
-        when(addressRepository.findAnyByUserIdAndAddressStatusNotDeleted(user.getId()))
-            .thenReturn(Optional.of(secondAddress));
-        doReturn(new OrderWithAddressesResponseDto()).when(addressService).findAllAddressesForCurrentOrder(uuid);
-
-        addressService.deleteCurrentAddressForOrder(firstAddressId, uuid);
-
-        Assertions.assertFalse(firstAddress.getActual());
-        assertEquals(AddressStatus.DELETED, firstAddress.getAddressStatus());
-        Assertions.assertTrue(secondAddress.getActual());
-
-        verify(addressRepository).findById(firstAddressId);
-        verify(addressRepository).findAnyByUserIdAndAddressStatusNotDeleted(user.getId());
-        verify(addressService).findAllAddressesForCurrentOrder(uuid);
-    }
-
-    @Test
-    void testDeleteCurrentAddressForOrderWhenAddressIsNotActual() {
-        UBSClientServiceImpl ubsClientServiceSpy = spy(ubsService);
-
-        Long firstAddressId = 1L;
-        Address firstAddress = getAddress();
-        firstAddress.setId(firstAddressId);
-        User user = getUser();
-        firstAddress.setUser(user);
-        String uuid = user.getUuid();
-
-        when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
-
-        doReturn(new OrderWithAddressesResponseDto()).when(addressService).findAllAddressesForCurrentOrder(uuid);
-
-        addressService.deleteCurrentAddressForOrder(firstAddressId, uuid);
-
-        assertEquals(AddressStatus.DELETED, firstAddress.getAddressStatus());
-
-        verify(addressRepository).findById(firstAddressId);
-        verify(addressRepository, times(0)).findAnyByUserIdAndAddressStatusNotDeleted(anyLong());
-        verify(addressService).findAllAddressesForCurrentOrder(uuid);
-    }
-
-    @Test
-    void testDeleteCurrentAddressForOrderWithUnexistingAddress() {
-        UBSClientServiceImpl ubsClientServiceSpy = spy(ubsService);
-
-        Long addressId = 1L;
-
-        when(addressRepository.findById(addressId)).thenReturn(Optional.empty());
-
-        NotFoundException exception = assertThrows(NotFoundException.class,
-            () -> addressService.deleteCurrentAddressForOrder(addressId, "qwe"));
-
-        assertEquals(NOT_FOUND_ADDRESS_ID_FOR_CURRENT_USER + addressId, exception.getMessage());
-
-        verify(addressRepository).findById(addressId);
-        verify(addressRepository, times(0)).findAnyByUserIdAndAddressStatusNotDeleted(anyLong());
-        verify(addressService, times(0)).findAllAddressesForCurrentOrder(anyString());
-    }
-
-    @Test
-    void testDeleteCurrentAddressForOrderForWrongUser() {
-        UBSClientServiceImpl ubsClientServiceSpy = spy(ubsService);
-
-        Long firstAddressId = 1L;
-        Address firstAddress = getAddress();
-        firstAddress.setId(firstAddressId);
-        User user = getUser();
-        firstAddress.setUser(user);
-        String uuid = "qwe";
-
-        when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
-
-        AccessDeniedException exception = assertThrows(AccessDeniedException.class,
-            () -> addressService.deleteCurrentAddressForOrder(firstAddressId, uuid));
-
-        assertEquals(CANNOT_DELETE_ADDRESS, exception.getMessage());
-
-        verify(addressRepository).findById(firstAddressId);
-        verify(addressRepository, times(0)).findAnyByUserIdAndAddressStatusNotDeleted(anyLong());
-        verify(addressService, times(0)).findAllAddressesForCurrentOrder(anyString());
-    }
-
-    @Test
-    void testDeleteCurrentAddressForOrderWhenAddressAlreadyDeleted() {
-        UBSClientServiceImpl ubsClientServiceSpy = spy(ubsService);
-
-        Long firstAddressId = 1L;
-        Address firstAddress = getAddress();
-        firstAddress.setId(firstAddressId);
-        firstAddress.setAddressStatus(AddressStatus.DELETED);
-        User user = getUser();
-        firstAddress.setUser(user);
-        String uuid = user.getUuid();
-
-        when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
-
-        BadRequestException exception = assertThrows(BadRequestException.class,
-            () -> addressService.deleteCurrentAddressForOrder(firstAddressId, uuid));
-
-        assertEquals(CANNOT_DELETE_ALREADY_DELETED_ADDRESS, exception.getMessage());
-
-        verify(addressRepository).findById(firstAddressId);
-        verify(addressRepository, times(0)).findAnyByUserIdAndAddressStatusNotDeleted(anyLong());
-        verify(addressService, times(0)).findAllAddressesForCurrentOrder(anyString());
-    }
-
-    @Test
-    void testDeleteCurrentAddressForOrderWhenItIsLastAddress() {
-        UBSClientServiceImpl ubsClientServiceSpy = spy(ubsService);
-
-        Long firstAddressId = 1L;
-        Address firstAddress = getAddress();
-        firstAddress.setId(firstAddressId);
-        firstAddress.setActual(true);
-        User user = getUser();
-        firstAddress.setUser(user);
-        String uuid = user.getUuid();
-
-        when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
-        when(addressRepository.findAnyByUserIdAndAddressStatusNotDeleted(user.getId())).thenReturn(Optional.empty());
-        doReturn(new OrderWithAddressesResponseDto()).when(addressService).findAllAddressesForCurrentOrder(uuid);
-
-        addressService.deleteCurrentAddressForOrder(firstAddressId, uuid);
-
-        Assertions.assertFalse(firstAddress.getActual());
-        assertEquals(AddressStatus.DELETED, firstAddress.getAddressStatus());
-
-        verify(addressRepository).findById(firstAddressId);
-        verify(addressRepository).findAnyByUserIdAndAddressStatusNotDeleted(user.getId());
-        verify(addressService).findAllAddressesForCurrentOrder(uuid);
-    }
-
-    @Test
-    void testMakeAddressActual() {
-        Long firstAddressId = 1L;
-        Long secondAddressId = 2L;
-        Address firstAddress = getAddress();
-        firstAddress.setId(firstAddressId);
-        Address secondAddress = getAddress();
-        secondAddress.setId(secondAddressId);
-        secondAddress.setActual(true);
-        User user = getUser();
-        firstAddress.setUser(user);
-        String uuid = user.getUuid();
-
-        when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
-        when(addressRepository.findByUserIdAndActualTrue(user.getId())).thenReturn(Optional.of(secondAddress));
-
-        addressService.makeAddressActual(firstAddressId, uuid);
-
-        Assertions.assertTrue(firstAddress.getActual());
-        assertFalse(secondAddress.getActual());
-
-        verify(addressRepository).findById(firstAddressId);
-        verify(addressRepository).findByUserIdAndActualTrue(user.getId());
-        verify(modelMapper).map(firstAddress, AddressDto.class);
-    }
-
-    @Test
-    void testMakeAddressActualWhereUserNotHaveActualAddress() {
-        Long firstAddressId = 1L;
-        Address firstAddress = getAddress();
-        firstAddress.setId(firstAddressId);
-        User user = getUser();
-        firstAddress.setUser(user);
-        String uuid = user.getUuid();
-
-        when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
-        when(addressRepository.findByUserIdAndActualTrue(user.getId())).thenReturn(Optional.empty());
-
-        NotFoundException exception = assertThrows(NotFoundException.class,
-            () -> addressService.makeAddressActual(firstAddressId, uuid));
-
-        assertEquals(ACTUAL_ADDRESS_NOT_FOUND, exception.getMessage());
-
-        verify(addressRepository).findById(firstAddressId);
-        verify(addressRepository).findByUserIdAndActualTrue(user.getId());
-        verify(modelMapper, times(0)).map(any(), any());
-    }
-
-    @Test
     void testMakeAddressActualWhenAddressIsAlreadyActual() {
         Long firstAddressId = 1L;
         User user = getUser();
@@ -2640,72 +2084,6 @@ class UBSClientServiceImplTest {
         verify(addressRepository).findById(firstAddressId);
         verify(addressRepository, times(0)).findByUserIdAndActualTrue(anyLong());
         verify(modelMapper).map(firstAddress, AddressDto.class);
-    }
-
-    @Test
-    void testMakeAddressActualWhenAddressNotFound() {
-        Long firstAddressId = 1L;
-        User user = getUser();
-        String uuid = user.getUuid();
-
-        when(addressRepository.findById(firstAddressId)).thenReturn(Optional.empty());
-
-        NotFoundException exception = assertThrows(NotFoundException.class,
-            () -> addressService.makeAddressActual(firstAddressId, uuid));
-
-        assertEquals(NOT_FOUND_ADDRESS_ID_FOR_CURRENT_USER + firstAddressId, exception.getMessage());
-
-        verify(addressRepository).findById(firstAddressId);
-        verify(addressRepository, times(0)).findByUserIdAndActualTrue(anyLong());
-        verify(modelMapper, times(0)).map(any(), any());
-    }
-
-    @Test
-    void testMakeAddressActualWhenAddressNotBelongsToUser() {
-        Long firstAddressId = 1L;
-        Long userId = 2L;
-        User user = getUser();
-        user.setId(userId);
-        String uuid = user.getUuid();
-        Address firstAddress = getAddress();
-        User addressOwner = getUser();
-        addressOwner.setUuid("randomUuid");
-        firstAddress.setId(firstAddressId);
-        firstAddress.setUser(addressOwner);
-
-        when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
-
-        AccessDeniedException exception = assertThrows(AccessDeniedException.class,
-            () -> addressService.makeAddressActual(firstAddressId, uuid));
-
-        assertEquals(CANNOT_ACCESS_PERSONAL_INFO, exception.getMessage());
-
-        verify(addressRepository).findById(firstAddressId);
-        verify(addressRepository, times(0)).findByUserIdAndActualTrue(anyLong());
-        verify(modelMapper, times(0)).map(any(), any());
-    }
-
-    @Test
-    void testMakeAddressActualWhenAddressIdDeleted() {
-        Long firstAddressId = 1L;
-        User user = getUser();
-        String uuid = user.getUuid();
-        Address firstAddress = getAddress();
-        firstAddress.setId(firstAddressId);
-        firstAddress.setUser(user);
-        firstAddress.setAddressStatus(AddressStatus.DELETED);
-        user.setAddresses(List.of(firstAddress));
-
-        when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
-
-        BadRequestException exception = assertThrows(BadRequestException.class,
-            () -> addressService.makeAddressActual(firstAddressId, uuid));
-
-        assertEquals(CANNOT_MAKE_ACTUAL_DELETED_ADDRESS, exception.getMessage());
-
-        verify(addressRepository).findById(firstAddressId);
-        verify(addressRepository, times(0)).findByUserIdAndActualTrue(anyLong());
-        verify(modelMapper, times(0)).map(any(), any());
     }
 
     @Test
@@ -4341,136 +3719,5 @@ class UBSClientServiceImplTest {
         verify(regionRepository).findRegionByNameEnOrNameUk(anyString(), anyString());
         verify(cityRepository).findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString());
         verify(districtRepository).findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString());
-    }
-
-    @Test
-    void getAllDistrictsForKyivTest() {
-        when(districtRepository.findAllByCityId(anyLong())).thenReturn(List.of(getDistrict()));
-
-        addressService.getAllDistrictsForKyiv();
-
-        verify(districtRepository).findAllByCityId(anyLong());
-    }
-
-    @Test
-    void updateCurrentAddressForOrderWithInvalidUserTest() {
-        when(userRepository.findByUuid(anyString())).thenReturn(null);
-
-        assertThrows(NotFoundException.class,
-            () -> addressService.updateCurrentAddressForOrder(null, USER_UUID));
-
-        verify(userRepository).findByUuid(anyString());
-    }
-
-    @Test
-    void updateCurrentAddressForOrderWithExistingDeletedAddressTest() {
-        OrderAddressDtoRequest dtoRequest = getTestOrderAddressDtoRequest2();
-        User user = getUser();
-        Address address = getAddress(1L)
-            .setUser(user)
-            .setAddressStatus(AddressStatus.DELETED);
-        CreateAddressRequestDto dto = getAddressRequestDto2();
-
-        when(userRepository.findByUuid(anyString())).thenReturn(user);
-        when(addressRepository.findById(anyLong())).thenReturn(Optional.of(address));
-        when(addressRepository.findAllByUserId(anyLong())).thenReturn(List.of(address));
-        when(modelMapper.map(any(), eq(CreateAddressRequestDto.class))).thenReturn(dto);
-
-        addressService.updateCurrentAddressForOrder(dtoRequest, USER_UUID);
-
-        verify(userRepository, times(2)).findByUuid(anyString());
-        verify(addressRepository).findById(anyLong());
-        verify(addressRepository).findAllByUserId(anyLong());
-        verify(modelMapper).map(any(), eq(CreateAddressRequestDto.class));
-        verify(addressRepository, times(2)).save(any());
-    }
-
-    @Test
-    void updateCurrentAddressForOrderWithNoExistingRegionTest() {
-        OrderAddressDtoRequest dtoRequest = getTestOrderAddressDtoRequest2();
-        User user = getUser();
-        Address address = getAddress(1L)
-            .setUser(user)
-            .setAddressStatus(AddressStatus.DELETED);
-        CreateAddressRequestDto dto = getAddressRequestDto();
-
-        when(userRepository.findByUuid(anyString())).thenReturn(user);
-        when(addressRepository.findById(anyLong())).thenReturn(Optional.of(address));
-        when(addressRepository.findAllByUserId(anyLong())).thenReturn(List.of(address));
-        when(modelMapper.map(any(), eq(CreateAddressRequestDto.class))).thenReturn(dto);
-        when(modelMapper.map(any(), eq(Address.class))).thenReturn(getAddress());
-        when(regionRepository.findRegionByNameEnOrNameUk(anyString(), anyString()))
-            .thenReturn(Optional.empty());
-
-        assertThrows(BadRequestException.class,
-            () -> addressService.updateCurrentAddressForOrder(dtoRequest, USER_UUID));
-
-        verify(userRepository).findByUuid(anyString());
-        verify(addressRepository).findById(anyLong());
-        verify(addressRepository).findAllByUserId(anyLong());
-        verify(modelMapper).map(any(), eq(CreateAddressRequestDto.class));
-        verify(modelMapper).map(any(), eq(Address.class));
-        verify(regionRepository).findRegionByNameEnOrNameUk(anyString(), anyString());
-    }
-
-    @Test
-    void testAreAddressesEqual() throws Exception {
-        Method areAddressesEqualMethod = UBSClientServiceImpl.class.getDeclaredMethod("areAddressesEqual",
-            CreateAddressRequestDto.class, CreateAddressRequestDto.class);
-        areAddressesEqualMethod.setAccessible(true);
-
-        CreateAddressRequestDto address1 = getAddressRequestDtoReflection();
-        CreateAddressRequestDto address2 = getAddressRequestDtoReflection2();
-        CreateAddressRequestDto address3 = getAddressRequestDtoReflection3();
-        CreateAddressRequestDto address4 = getAddressRequestDtoReflection4();
-        CreateAddressRequestDto address5 = getAddressRequestDtoReflection5();
-
-        boolean result1 = (boolean) areAddressesEqualMethod.invoke(ubsClientService, address1, address2);
-        assertTrue(result1);
-
-        boolean result2 = (boolean) areAddressesEqualMethod.invoke(ubsClientService, address1, address3);
-        assertFalse(result2);
-
-        boolean result3 = (boolean) areAddressesEqualMethod.invoke(ubsClientService, address1, null);
-        assertFalse(result3);
-
-        boolean result4 = (boolean) areAddressesEqualMethod.invoke(ubsClientService, null, null);
-        assertFalse(result4);
-
-        boolean result5 = (boolean) areAddressesEqualMethod.invoke(ubsClientService, address1, address4);
-        assertFalse(result5);
-
-        boolean result6 = (boolean) areAddressesEqualMethod.invoke(ubsClientService, address1, address5);
-        assertTrue(result6);
-    }
-
-    @Test
-    void updateCurrentAddressForOrderWithNoExistingAddressTest() throws Exception {
-        Method setLocations = UBSClientServiceImpl.class.getDeclaredMethod("setLocations",
-            CreateAddressRequestDto.class, Address.class);
-        setLocations.setAccessible(true);
-
-        Address address = getAddress();
-        CreateAddressRequestDto dto = getAddressRequestToSaveDto();
-
-        when(regionRepository.findRegionByNameEnOrNameUk(anyString(), anyString()))
-            .thenReturn(Optional.of(getRegion()));
-        when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
-            .thenReturn(Optional.empty());
-        when(addressMapper.convert(any(), eq(City.class))).thenReturn(getCity());
-        when(cityRepository.save(any())).thenReturn(getCity());
-        when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
-            .thenReturn(Optional.empty());
-        when(addressMapper.convert(any(), eq(District.class))).thenReturn(getDistrict());
-
-        setLocations.invoke(ubsClientService, dto, address);
-
-        verify(regionRepository).findRegionByNameEnOrNameUk(anyString(), anyString());
-        verify(cityRepository).findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString());
-        verify(addressMapper).convert(any(), eq(City.class));
-        verify(districtRepository).findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString());
-        verify(addressMapper).convert(any(), eq(District.class));
-        verify(cityRepository).save(any());
-        verify(districtRepository).save(any());
     }
 }
