+++ service/src/test/java/greencity/service/ubs/UBSManagementServiceImplTest.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import greencity.ModelUtils;
 import greencity.client.UserRemoteClient;
 import greencity.constant.OrderHistory;
-import greencity.dto.address.UpdateAddressDto;
 import greencity.dto.bag.AdditionalBagInfoDto;
 import greencity.dto.bag.BagInfoDto;
 import greencity.dto.bag.BagMappingDto;
@@ -19,14 +18,11 @@ import greencity.dto.order.EcoNumberDto;
 import greencity.dto.order.ExportDetailsDto;
 import greencity.dto.order.ExportDetailsDtoUpdate;
 import greencity.dto.order.NotTakenOrderReasonDto;
-import greencity.dto.order.OrderAddressDtoResponse;
-import greencity.dto.order.OrderAddressExportDetailsDtoUpdate;
 import greencity.dto.order.OrderCancellationReasonDto;
 import greencity.dto.order.OrderDetailInfoDto;
 import greencity.dto.order.OrderDetailStatusDto;
 import greencity.dto.order.OrderDetailStatusRequestDto;
 import greencity.dto.order.OrderInfoDto;
-import greencity.dto.order.ReadAddressByOrderDto;
 import greencity.dto.order.UpdateAllOrderPageDto;
 import greencity.dto.order.UpdateOrderPageAdminDto;
 import greencity.dto.pageble.PageableDto;
@@ -44,7 +40,6 @@ import greencity.entity.user.employee.Employee;
 import greencity.entity.user.employee.EmployeeOrderPosition;
 import greencity.entity.user.employee.Position;
 import greencity.entity.user.employee.ReceivingStation;
-import greencity.entity.user.ubs.OrderAddress;
 import greencity.enums.NotificationType;
 import greencity.enums.OrderPaymentStatus;
 import greencity.enums.OrderStatus;
@@ -119,8 +114,6 @@ import static greencity.ModelUtils.TEST_BAG_LIST;
 import static greencity.ModelUtils.TEST_BAG_MAPPING_DTO_LIST;
 import static greencity.ModelUtils.TEST_MAP_ADDITIONAL_BAG_LIST;
 import static greencity.ModelUtils.TEST_ORDER;
-import static greencity.ModelUtils.TEST_ORDER_ADDRESS_DTO_RESPONSE;
-import static greencity.ModelUtils.TEST_ORDER_ADDRESS_DTO_UPDATE;
 import static greencity.ModelUtils.TEST_ORDER_DETAILS_INFO_DTO_LIST;
 import static greencity.ModelUtils.TEST_PAYMENT_LIST;
 import static greencity.ModelUtils.TEST_USER;
@@ -136,10 +129,8 @@ import static greencity.ModelUtils.getExportDetailsRequest;
 import static greencity.ModelUtils.getExportDetailsRequestToday;
 import static greencity.ModelUtils.getFormedOrder;
 import static greencity.ModelUtils.getInfoPayment;
-import static greencity.ModelUtils.getLocation;
 import static greencity.ModelUtils.getNotificationParameterSet;
 import static greencity.ModelUtils.getOrder;
-import static greencity.ModelUtils.getOrderAddress;
 import static greencity.ModelUtils.getOrderBag;
 import static greencity.ModelUtils.getOrderDoneByUser;
 import static greencity.ModelUtils.getOrderExportDetails;
@@ -173,7 +164,6 @@ import static greencity.ModelUtils.getTariffsInfo;
 import static greencity.ModelUtils.getTestDetailsOrderInfoDto;
 import static greencity.ModelUtils.getTestOrderDetailStatusRequestDto;
 import static greencity.ModelUtils.getTestUser;
-import static greencity.ModelUtils.getUpdateAddressDto;
 import static greencity.ModelUtils.getUserNotificationForUnpaidOrder;
 import static greencity.ModelUtils.updateAllOrderPageDto;
 import static greencity.ModelUtils.updateOrderPageAdminDto;
@@ -202,9 +192,6 @@ class UBSManagementServiceImplTest {
 
     @Mock
     private FileService fileService;
-
-    @Mock
-    private AddressService addressService;
 
     @Mock(strictness = Mock.Strictness.LENIENT)
     OrderRepository orderRepository;
@@ -311,22 +298,6 @@ class UBSManagementServiceImplTest {
         PageableDto<CertificateDtoForSearching> actual =
             ubsManagementService.getAllCertificates(pageable, "points", SortingOrder.DESC);
         assertEquals(certificateDtoForSearchingPageableDto, actual);
-    }
-
-    @Test
-    void checkOrderNotFound() {
-        assertThrows(NotFoundException.class,
-            () -> addressService.getAddressByOrderId(10000000L));
-    }
-
-    @Test
-    void getAddressByOrderId() {
-        Order order = getOrder();
-        ReadAddressByOrderDto readAddressByOrderDto = ModelUtils.getReadAddressByOrderDto();
-        when(orderRepository.findById(1L)).thenReturn(Optional.of(order));
-        when(orderAddressRepository.findByOrderId(anyLong())).thenReturn(Optional.of(getOrderAddress()));
-        when(addressService.getAddressByOrderId(1L)).thenReturn(readAddressByOrderDto);
-        Assertions.assertNotNull(order);
     }
 
     @Test
@@ -601,40 +572,6 @@ class UBSManagementServiceImplTest {
         verify(orderRepository).findById(anyLong());
         verify(employeeRepository).findByEmail("test@gmail.com");
         verify(tariffsInfoRepository, atLeastOnce()).findTariffsInfoByIdForEmployee(anyLong(), anyLong());
-    }
-
-    @Test
-    void testUpdateAddress() {
-        Order order = getOrder();
-        OrderAddress orderAddress = getOrderAddress();
-        orderAddress.setId(1L);
-        orderAddress.setLocation(getLocation());
-        OrderAddressExportDetailsDtoUpdate dtoUpdate = ModelUtils.getOrderAddressExportDetailsDtoUpdate();
-        OrderAddress updatedOrderAddress = getOrderAddress();
-        updatedOrderAddress.setCity("Updated");
-
-        when(orderAddressRepository.findById(dtoUpdate.getId())).thenReturn(Optional.of(orderAddress));
-        when(addressService.updateOrderAddress(any())).thenReturn(updatedOrderAddress);
-        when(orderAddressRepository.save(orderAddress)).thenReturn(updatedOrderAddress);
-        when(modelMapper.map(updatedOrderAddress, OrderAddressDtoResponse.class))
-            .thenReturn(TEST_ORDER_ADDRESS_DTO_RESPONSE);
-        Optional<OrderAddressDtoResponse> actual =
-                addressService.updateAddress(TEST_ORDER_ADDRESS_DTO_UPDATE, order, "test@gmail.com");
-        assertEquals(Optional.of(TEST_ORDER_ADDRESS_DTO_RESPONSE), actual);
-
-        verify(orderAddressRepository).findById(dtoUpdate.getId());
-        verify(orderAddressRepository).save(updatedOrderAddress);
-        verify(eventService).saveEvent(OrderHistory.WASTE_REMOVAL_ADDRESS_CHANGE, "test@gmail.com", order);
-        verify(modelMapper).map(updatedOrderAddress, OrderAddressDtoResponse.class);
-        assertEquals(1L, updatedOrderAddress.getLocation().getId());
-        assertEquals(1L, updatedOrderAddress.getId());
-    }
-
-    @Test
-    void testUpdateAddressThrowsNotFoundOrderAddressException() {
-        Order order = getOrder();
-        assertThrows(NotFoundException.class,
-            () -> addressService.updateAddress(TEST_ORDER_ADDRESS_DTO_UPDATE, order, "abc"));
     }
 
     @Test
@@ -2435,42 +2372,5 @@ class UBSManagementServiceImplTest {
         verify(tariffsInfoRepository, atLeastOnce()).findTariffsInfoByIdForEmployee(anyLong(), anyLong());
         verify(orderStatusTranslationRepository).findAllBy();
         verify(orderPaymentStatusTranslationRepository).getAllBy();
-    }
-
-    @Test
-    void updateAddressTest() {
-        UpdateAddressDto updateAddressDto = getUpdateAddressDto();
-        String email = "test@email.com";
-        OrderAddressDtoResponse response = OrderAddressDtoResponse.builder()
-            .entranceNumber("1")
-            .build();
-        OrderAddress orderAddress = getOrderAddress();
-
-        when(orderRepository.findById(updateAddressDto.getOrderId())).thenReturn(Optional.of(getOrder()));
-        when(orderAddressRepository.findById(anyLong())).thenReturn(Optional.of(orderAddress));
-        when(addressService.updateOrderAddress(any(OrderAddressExportDetailsDtoUpdate.class)))
-            .thenReturn(orderAddress);
-        when(modelMapper.map(orderAddress, OrderAddressDtoResponse.class)).thenReturn(response);
-
-        addressService.addressUpdate(updateAddressDto, email);
-
-        verify(orderRepository).findById(updateAddressDto.getOrderId());
-        verify(orderAddressRepository).findById(anyLong());
-        verify(addressService).updateOrderAddress(any(OrderAddressExportDetailsDtoUpdate.class));
-        verify(modelMapper).map(orderAddress, OrderAddressDtoResponse.class);
-        verify(orderAddressRepository).save(any(OrderAddress.class));
-        verify(eventService).saveEvent(anyString(), anyString(), any(Order.class));
-    }
-
-    @Test
-    void updateAddressTestIfOrderNotFoundTest() {
-        UpdateAddressDto updateAddressDto = getUpdateAddressDto();
-        String email = "test@email.com";
-
-        when(orderRepository.findById(updateAddressDto.getOrderId())).thenReturn(Optional.empty());
-
-        assertThrows(NotFoundException.class, () -> addressService.addressUpdate(updateAddressDto, email));
-
-        verify(orderRepository).findById(updateAddressDto.getOrderId());
     }
 }
